move build_desc_chain to MockSplitQueue
Moved a helper function that creates a descriptor chain
from virtio-block/request.rs to the mock framework, so
that it can be used by multiple crates.
Also made it a bit more generic, and moved the creation
of the queue outside of the function. This way it is also
more obvious that there's a queue involved in building the
descriptor chain, as opposed to how things were before.
